requirements.txt
Django>=4.2.0,<5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.1.0
pillow>=10.0.0
pandas>=2.0.0
xlsxwriter>=3.1.0
reportlab>=4.0.0 